Restructure src into module-based api and front pages
