+++ src/front/pages/auth/ForcePasswordChange.jsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from "react";
-import { useGlobalReducer } from "../hooks/useGlobalReducer";
+import { useGlobalReducer } from "../../hooks/useGlobalReducer";
 import { useNavigate } from "react-router-dom";
 import toast from "react-hot-toast";
