+++ src/front/pages/auth/RegistrationLanding.jsx
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import { useGlobalReducer } from "../hooks/useGlobalReducer";
+import { useGlobalReducer } from "../../hooks/useGlobalReducer";
 import { useNavigate } from "react-router-dom";
 import toast from "react-hot-toast";
