+++ src/front/pages/public/Layout.jsx
@@ -1,9 +1,9 @@
 import React from "react";
 import { Outlet } from "react-router-dom";
 import { Toaster } from "react-hot-toast";
-import ScrollToTop from "../components/ScrollToTop";
-import { Navbar } from "../components/Navbar";
-import { Footer } from "../components/Footer";
+import ScrollToTop from "../../components/ScrollToTop";
+import { Navbar } from "../../components/Navbar";
+import { Footer } from "../../components/Footer";
 
 // Base component que mantiene el navbar y footer en todas las páginas.
 export const Layout = () => {
