+++ src/front/pages/public/Single.jsx
@@ -1,8 +1,8 @@
 // Import necessary hooks and components from react-router-dom and other libraries.
 import { Link, useParams } from "react-router-dom";  // To use link for navigation and useParams to get URL parameters
 import PropTypes from "prop-types";  // To define prop types for this component
-import rigoImageUrl from "../assets/img/rigo-baby.jpg"  // Import an image asset
-import useGlobalReducer from "../hooks/useGlobalReducer";  // Import a custom hook for accessing the global state
+import rigoImageUrl from "../../assets/img/rigo-baby.jpg"  // Import an image asset
+import useGlobalReducer from "../../hooks/useGlobalReducer";  // Import a custom hook for accessing the global state
 
 // Define and export the Single component which displays individual item details.
 export const Single = props => {
